requirements.txt
# Forcing torch cpu
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu


# ---------------------------
# Core API framework
# ---------------------------
fastapi==0.110.0          # High-performance Python web framework for APIs
uvicorn[standard]==0.27.1 # ASGI server to run FastAPI


# ---------------------------
# Database & ORM
# ---------------------------
sqlalchemy==2.0.27        # ORM for database access
alembic==1.13.1           # Database migrations
pydantic==2.6.1           # Data validation and serialization

# ---------------------------
# Configuration & utilities
# ---------------------------
python-dotenv==1.0.1      # Load environment variables from .env files


# ---------------------------
# ML baseline
# ---------------------------
scikit-learn==1.4.1.post1   # TF-IDF + Logistic Regression pipeline
joblib==1.3.2               # Persist/load ML models
pandas==2.2.1               # Optional: dataset handling

# ---------------------------
# RAG (local)
# ---------------------------
langchain==0.1.13
langchain-community==0.0.29
faiss-cpu==1.7.4
pypdf==4.1.0
sentence-transformers==2.6.1


# ---------------------------
# OpenaI
# ---------------------------
openai==1.40.0
httpx==0.26.0
httpcore==1.0.5


# ---------------------------
# UI
# ---------------------------
streamlit==1.35.0
requests==2.31.0



# ---------------------------
# Auth-users
# ---------------------------
argon2-cffi==23.1.0        # Password hashing with Argon2
python-jose==3.3.0         # JWT tokens
python-multipart==0.0.9    # Form handling if needed later
email-validator==2.1.1     # Required by Pydantic EmailStr